Move more tests out of `electron-browser` (#171508) (#175445)
* debt - move more tests out of `electron-browser` (#171508)

* move more tests

* move more tests

* skip test on windows where tilde is unsupported
