+++ src/vs/workbench/contrib/tags/common/workspaceTags.ts
@@ -6,6 +6,7 @@
 import { WorkbenchState, IWorkspace } from 'vs/platform/workspace/common/workspace';
 import { createDecorator } from 'vs/platform/instantiation/common/instantiation';
 import { URI } from 'vs/base/common/uri';
+import { getRemotes } from 'vs/platform/extensionManagement/common/configRemotes';
 
 export type Tags = { [index: string]: boolean | number | string | undefined };
 
@@ -23,4 +24,8 @@ export interface IWorkspaceTagsService {
 	getTelemetryWorkspaceId(workspace: IWorkspace, state: WorkbenchState): Promise<string | undefined>;
 
 	getHashedRemotesFromUri(workspaceUri: URI, stripEndingDotGit?: boolean): Promise<string[]>;
+}
+
+export async function getHashedRemotesFromConfig(text: string, stripEndingDotGit: boolean = false, sha1Hex: (str: string) => Promise<string>): Promise<string[]> {
+	return Promise.all(getRemotes(text, stripEndingDotGit).map(remote => sha1Hex(remote)));
 }
