+++ src/vs/workbench/contrib/themes/test/node/colorRegistryExport.test.ts
